added FriendRequestController completing refactoring suggestion in #82
